Handle correspondence games with little time left (#594)
During correspondence games, the time left on the clock is generally ignored since the latter is measured in days. However, if lichess-bot is started when a correspondence game clock is about to expire, the delay caused by queuing plus the move time in the user's config file can result in a bot losing on time. This is fixed in two parts:

- On startup: if it is the bot's move and the time on the bot's clock is less than ten times the sum of the correspondence: checkin_period and correspondence: move_time, then start the game immediately instead of queuing.
- In a game: if the time left on the bot's clock is less than the correspondence: move_time in the user's config file, use the game clock.


* Do not queue correspondence games if low on time

* Do not let clock run out in correspondence games

If the time left on the clock is less than the search time in the user's
config file, use the clock time to pick a move.

* Handle games with little time separately

Correspondence games with little time left are put in a separate list.
Whenever there is room (as defined by concurrency) start the game
with the least amount of time left.

* Fix for getting game ID

With games sorted by time remaining, choose the first game since that one
has the least amount of time remaining.

Since the entire game JSON is stored, pick out the ID to send to lichess to
start the game.

* Fix detection of correspondence games

The test game["perf"] == "correspondence" only detects standard chess
correspondence games, not variates or From Position. Changing this
to game["speed"] == "correspondence" fixes this.