Track active game IDs (#598)
I believe many bug reports related to lichess-bot clients no longer issuing or accepting challenges is due to the counts of active games not being accurate. For example, when a game ends with a threefold draw, lichess sends a game stream event that indicates the bot should keep playing. When the bot tries to play a move, an exception is thrown because the game is over on the lichess server (I've submitted an issue). This exception ends the play_game() thread without creating the local_game_done event. This causes the busy_processes count to not be decremented. Eventually, after enough of these errors, busy_processes will equal max_games and no more games will be played until lichess-bot is restarted. Since threefold repetition draws are relatively rare, this would explain why the bug takes so long to manifest.

This PR makes two changes:
1. Keep track of the IDs of active games instead of just the counts. An active game is one that is currently being played or one for which the client bot has accepted a challenge. This will help with bug reporting since it will reveal when too many stale games are preventing bots from playing more games--as is the case for the threefold draw bug above.
2. Remove games from this list when either a local_game_done or a gameFinish event is received. This fixes at least one cause of lichess-bot stopping all new games.


* Replace busy/queued_processes with game id list

Instead of maintaining a count of games, keep a list of game IDs which
are either being played or the bot has accepted a challenge. This makes
debugging easier in cases where the counts get out of sync with the
lichess game server.

* Nicer game ID output

* Discard game from gameFinish event

Fix local_game_done event so that the id location matches the
gameFinish event.