Forcing defaults for some falsey values
For the engine:working_dir configuration, either a missing paramter or
an empty string should result in the default value of os.getcwd() being
used. This is now enforced with a new parameter on set_config_default().

The insertion of default values has been moved to a separate function so
that it can be imported and called from the test framework.