Simplify correspondence game handling
Instead of putting special entries in the correspondence queue and
interpreting them, keep a count of how many games should be
started each time the correspondence check-in function in called.

After a correspondence ping, this count is set to the number of game IDs
in the queue. After a local_game_done, this number is equal to the number
of games leftover after the last check-in.

This slightly changes the behavior of correspondence handling. After a
correspondence_ping, all correspondence games are checked, even those
that were already recently checked. Now, the checkin_period is a maximum
time between game checkins, whereas before, this was a minimum.