Fix list configurations with empty values in file
If a user has a configuration file like this:

matchmaking:
  challenge_increment:
  challenge_days:

These will be translated with the config default functions into

matchmaking:
  challenge_increment:
    - null
  challenge_days:
    - null

This commit corrects this so that null entries result in empty lists.

The force_falsey_values argument does not work because we want
a single entry of 0 to be valid.