Ignore errors when updating user profile
Since this is only used to update the engines rating when looking for
similarly rated opponents, missing one update is not critical since it
is repeated every 5 minutes.

The other call to Lichess.get_profile() in start_lichess_bot() is unchanged
since an error here stops the program before anything starts. This may be
an early warning that something is wrong with lichess.org. It is quick for the
user to restart the program.