Handle when getting PGN from lichess fails
In this case, a failure to download the PGN from lichess will return an
empty string. This will cause the PGN file to be written from local
information. The header of the PGN will be missing information
(perhaps this can be rectified in the future), but the moves and engine
output will be written. The file name will contain enough information
to manually download the game from lichess.