Slow down challenge creation as daily limit is approached (#679)
* Slow down challenges to keep under daily limit

Keep a record of challenges, each record expiring after a day. As
the number of challenges increases, slow down challenge creation
to keep under the API rate limit.

* Inform user about when next challenge is due

This should let users know that a slowdown in challenge creation
is deliberate.

* Show next challenge time after a decline

* Record challenge time before calling lichess

Make sure an exception doesn't cause an undercount of challenges.

* Log how many challenges have occurred in 24 hours

* Save challenge datetimes to file between session

Unless the challenge times are written to a file, the number of
challenges in the last 24 hours will be lost in a least two situations:
1. lichess-bot is stopped and restarted
2. The lichess server is restarted, which causes lichess_bot_main() to exit.

* Change log level of cancelled challenge message

* Fix loading of challenge times from file

Now, timestamps read from the file will be written back without
being changed.

Other changes: use double-quotes instead of single; include newline in timestamp_format to simplify reading/writing.

* Reset last_challenge_created before challenging

In case an exception occurs.

* Don't print exception for challenge rate limit

A challenge failing is not uncommon, especially if rate-limited.
Log the full error message to the debug log.