Replace timers with bools for challenge filtering
Challenge filters no longer expire after a set time. These filters
are not permanent, since they get reset when lichess-bot is
restarted or the lichess servers are rebooted.

Matchmaking members are renamed to match their function,
rather than the implementation.

The config value is renamed to challenge_filter. The value for
delay_after_decline is used if available for backwards compatibility.

Documentation is updated.