Update README.md (#711)
Due to conflicts with python modules that are installed from Linux package managers, some Python modules can't be installed through pip outside of a virtual environment. Due to PEP 668, attempting to use pip to install virtualenv will fail with a warning about the module being "externally managed." This change updates the README to change the instructions for installing virtualenv from pip to the OS package manager.