All loggig uses logging_queue
I believe this should fix the garbled logs on the console in Windows CMD and Mac terminals. Essentially, Once logging_listener_proc() is
active, switch logging in the main thread to use the logging_queue,
just like each game thread. This way, it's impossible for two log
messages to be written to the console simultaneously. The name
of the function that sets up the log queue was changed to reflect
its expanded function.

The one side effect is that shutting down lichess-bot with Ctrl-C seems
to take a few seconds now.