combine fake_think_time with time already consumed (#796)
* combine fake_think_time, rate_limiting_delay with time already consumed

There are 3 different delays:
1/ time spend by the engine
2/ rate_limiting_delay
3/ fake_think_time formula

The problem is each of these is unaware of the other 2, and we just add them all up. This is a
recipie for disaster. Instead we want to spend the max.

* Revert "combine fake_think_time, rate_limiting_delay with time already consumed"

This reverts commit 93851caab52f67af989463a3eef89d5169e774fb.

* do not add fake_think_time to what the engine has spent

Also, heed move_overhear in fake_think_time formula.

* rename fake_thinking() to fake_think_time()
better name, as the function does nothing, just returns the sleep value.

* fixes
- capitalize
- no need to cap sleep at remaining, it is smaller by construction