Reset logging config before exiting (#800)
Because the main thread logs to a queue that is then handled by
a listener process, terminating the listener process stops logging
for the rest of the program. This means that, if lichess-bot is stopped
by an exception, then the exception logging at the bottom of the file
(line 1029) does not do anything.

This change reverts the main thread back to direct logging before
terminating the logging queue thread.