Write PGN to files in main thread
Instead of writing the PGN files from the game thread/process, create
a string of the game record and send it to the main thread in the
local_game_done event. This way, even if the bot finishes multiple
games simultaneously and the bot's config specifies "all" for
pgn_file_grouping, only one game will be written to the file at a time.