Simplify PGN transport
All information needed to write the PGN file is in the local_game_done
event. The information is either loaded at the end of a game in the
final_queue_entries() function, or, in case of an error during the game,
in the game_error_handler() inside start_game_thread(). The
model.Game instance from inside the game thead is no longer needed.