Prevent hanging due to engine communication error
The python-chess SimpleEngine class is meant to be used as a context
manager, so make use of its __enter__() and __exit__() methods to
ensure complete startup and shutdown. Also, if the engine is shutting
down due to an exception, do not attempt any more communication
(i.e., ping and quit).