Create a test to show new timeout handling works
The new test uses a rudimentary engine (buggy_engine.py) that only plays
moves  to let black win with a scholars mate. However, on the third
move, it pauses long enough for the python-chess library to intervene
and raise an asyncio.TimeoutError exception. In a successful test, the
engine restarts and finishes the game.

Also, add an optional variable for a text log file to collect useful
data during tests. See testing_log_file_name on line 88.