Prevent hangs due to timeout errors
If a user sets a finite value for movetime in their config file
(e.g., engine: uci_options: go_commands: movetime) and the
engine takes significantly longer to pick a move (10 seconds
longer by default), then the python-chess framework will
cancel the search for a move and raise a TimeoutError.

Before this PR, this would cause lichess-bot to hang and stop
listening to the engine. Now, the error propagates to the
backoff decorator so that the engine can be restarted.