Limit simultaneous challenges from one user
In order to limit the number of games with a single opponent, the number
of queued challenges from a single opponent must be limited as well.
This change adds the number of active challenges to the number of active
games to keep the eventual number of simultaneous games with a single
user under the max_simultaneous_games_per_user configuration.