Remove tracking of daily challenges
The purpose of tracking daily challenges was to slow down the issuing of
challenges when the rate limit of 400/day was being approached. However,
now that the number of games is limited to 100/day (whether the bot
issues or receives the challenge), this tracking is redundant.

The rate at which challenges are issued is now entirely controlled by
the matchmaking: challenge_timeout configuration, which specifies how
long after a game ends to issue a new challenge.