Put daily game limit info in user messages
Put the rate-limit timeout into the RateLimitedError exception so other parts of the program can use it.

Add a rate-limit Timer to the Matchmaking class so that the user's logs aren't fill with rate-limited messages. Also, change the format of the "Next challenge will be created after" message to include the date since the delay may last into the next day.