Handle failed token test without leaking token (#1138)
If the line `token_info = token_response[token]` fails with a `KeyError`
(due to an error response from lichess), the user's token will be
printed in the error message. Use `token_response.get()` to prevent
this.

Also, use `.get()` on the resulting `token_info` to print a more useful
error message.