Use Timers instead of bools to filter opponents to allow temporary stops to challenges (#1141)
* Make matchmaking challenge filters time-limited

This will allow for filters to expire in case challenge declinations are
for temporary reasons. Also, in the case of an opponent reaching the
100-game daily limit, this will be useful for trying again in the time
given in the interval.

* Factor out variable

* Distinguish between bot limit and opponent limit

In an upcoming lichess update, challenging an opponent that has played 100 bot games will result in a similar error response to when the user's bot has played 100 bot games. This change distinguishes between them by checking if the error message contains the opponent's name.

This could also be distinguished by checking if the status code is 429 (user's bot) or 400 (opponent bot). But that information isn't returned from the Lichess.challenge() call.

* Be more specific about challenge pause delay

* Factor out function to get timeout from error

* Use status code to tell who is rate limited

A status code of 400 indicates that the opponent is rate limited, while
429 indicates that the bot is limited.

Also, put who is rate limited (if anyone) in the challenge response so
the Matchmaking class can use it.

* Simplify challenge control flow

* Only set rate limit when bot gets 429 status code

Don't set rate limit when opponent is rate limited (400 status code).

* Create a method to handle responses to challenges