Assume game is active until lichess says otherwise
Previously, if a communication error prevented lichess from sending
ongoing games, an empty list would be returned from the Lichess
function. This could cause game_is_active() to return False during a
temporary communication error during a game. This would cause a game to
be aborted locally even if the game is actually ongoing.

Now, when checking if a game is ongoing, don't assume a game is over
until a list is successfully received from lichess.